fixing wishlist to specific users
