+++ src/main/java/com/example/dreamwishes/controller/UserController.java
@@ -1,8 +1,6 @@
 package com.example.dreamwishes.controller;
 
 import com.example.dreamwishes.entity.Users;
-import com.example.dreamwishes.entity.Wishlist;
-import com.example.dreamwishes.repository.WishlistRepository;
 import com.example.dreamwishes.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,7 +15,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.servlet.http.HttpSession;
-import java.util.List;
 
 @Controller
 public class UserController {
@@ -27,8 +24,7 @@ public class UserController {
     @Autowired
     private UserService userService;
 
-    @Autowired
-    private WishlistRepository wishlistRepository;
+
 
     @GetMapping("/login")
     public String showLoginPage() {
@@ -49,31 +45,6 @@ public class UserController {
         } else {
             model.addAttribute("error", "Invalid username or password");
             return "login";
-        }
-    }
-@GetMapping("/wishes")
-public String showUserWishes(HttpSession session, Model model) {
-    Long userId = (Long) session.getAttribute("userId");
-    if (userId != null) {
-        Users currentUser = userService.getUserById(userId).orElse(null);
-        List<Wishlist> userWishes = wishlistRepository.findByUser(currentUser); // Assuming you have a method in your repository to get wishes by user
-        model.addAttribute("user", currentUser);
-        model.addAttribute("wishes", userWishes);
-        return "wishes"; // Return the wishes view
-    } else {
-        return "redirect:/login";
-    }
-}
-    @PostMapping("/wishes/add")
-    public String addWish(@ModelAttribute Wishlist newWish, HttpSession session) {
-        Long userId = (Long) session.getAttribute("userId");
-        if (userId != null) {
-            Users currentUser = userService.getUserById(userId).orElse(null);
-            newWish.setUser(currentUser); // Set the user of the new wish
-            wishlistRepository.save(newWish); // Save the new wish
-            return "redirect:/wishes"; // Redirect back to the wishes page
-        } else {
-            return "redirect:/login";
         }
     }
