+++ src/main/java/com/example/dreamwishes/controller/WishController.java
@@ -1,50 +1,58 @@
 package com.example.dreamwishes.controller;
 
+import com.example.dreamwishes.entity.Users;
 import com.example.dreamwishes.entity.Wishlist;
-import com.example.dreamwishes.service.WishlistService;
+import com.example.dreamwishes.repository.WishlistRepository;
+import com.example.dreamwishes.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import javax.servlet.http.HttpSession;
+import java.util.List;
+
 @Controller
 @RequestMapping("/wishes/add")
 public class WishController {
 
+    @Autowired
+    private UserService userService;
+    @Autowired
+    private WishlistRepository wishlistRepository;
 
-    @GetMapping
+    @GetMapping("/add")
     public String getAddWishPage() {
         return "addwish";
     }
 
-    private final WishlistService wishlistService;
-
-    public WishController(WishlistService wishlistService) {
-        this.wishlistService = wishlistService;
-    }
-
     @PostMapping("/add")
-    public String addWish(@ModelAttribute Wishlist wishlist) {
-        Wishlist createdWishlist = wishlistService.createWishlist(wishlist);
-        if (createdWishlist != null) {
-            return "redirect:/wishes"; // Redirect to the wishes page after successful creation
+    public String addWish(@ModelAttribute Wishlist newWish, HttpSession session) {
+        Long userId = (Long) session.getAttribute("userId");
+        if (userId != null) {
+            Users currentUser = userService.getUserById(userId).orElse(null);
+            newWish.setUser(currentUser); // Set the user of the new wish
+            wishlistRepository.save(newWish); // Save the new wish
+            return "redirect:/wishes"; // Redirect back to the wishes page
         } else {
-            return "error"; // Return an error page if the wish could not be created
+            return "redirect:/login";
         }
     }
 
-    //old one
-    /*@PostMapping("/add")
-    public ResponseEntity<Wishlist> addWish(@RequestBody Wishlist wishlist) {
-        Wishlist createdWishlist = wishlistService.createWishlist(wishlist);
-        if (createdWishlist != null) {
-            return ResponseEntity
-                    .created(URI.create("/api/wishlists/" + createdWishlist.getWishlistId()))
-                    .body(createdWishlist);
+    @GetMapping
+    public String showUserWishes(HttpSession session, Model model) {
+        if (session.getAttribute("loggedIn") != null && (Boolean) session.getAttribute("loggedIn")) {
+            Long userId = (Long) session.getAttribute("userId");
+            Users currentUser = userService.getUserById(userId).orElse(null);
+            List<Wishlist> userWishes = wishlistRepository.findByUser(currentUser);
+            model.addAttribute("user", currentUser);
+            model.addAttribute("wishes", userWishes);
+            return "wishes";
         } else {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+            return "redirect:/login";
         }
-
-     */
+    }
 }
